+++ macos/avfoundation/video_composition_layer_instruction.gen.go
@@ -22,7 +22,7 @@ type IVideoCompositionLayerInstruction interface {
 	objc.IObject
 	GetCropRectangleRampForTimeStartCropRectangleEndCropRectangleTimeRange(time coremedia.Time, startCropRectangle *coregraphics.Rect, endCropRectangle *coregraphics.Rect, timeRange *coremedia.TimeRange) bool
 	GetTransformRampForTimeStartTransformEndTransformTimeRange(time coremedia.Time, startTransform *coregraphics.AffineTransform, endTransform *coregraphics.AffineTransform, timeRange *coremedia.TimeRange) bool
-	GetOpacityRampForTimeStartOpacityEndOpacityTimeRange(time coremedia.Time, startOpacity *float64, endOpacity *float64, timeRange *coremedia.TimeRange) bool
+	GetOpacityRampForTimeStartOpacityEndOpacityTimeRange(time coremedia.Time, startOpacity *float32, endOpacity *float32, timeRange *coremedia.TimeRange) bool
 	TrackID() objc.Object
 }
 
@@ -78,7 +78,7 @@ func (v_ VideoCompositionLayerInstruction) GetTransformRampForTimeStartTransform
 // Obtains the opacity ramp that includes a specified time. [Full Topic]
 //
 // [Full Topic]: https://developer.apple.com/documentation/avfoundation/avvideocompositionlayerinstruction/1388471-getopacityrampfortime?language=objc
-func (v_ VideoCompositionLayerInstruction) GetOpacityRampForTimeStartOpacityEndOpacityTimeRange(time coremedia.Time, startOpacity *float64, endOpacity *float64, timeRange *coremedia.TimeRange) bool {
+func (v_ VideoCompositionLayerInstruction) GetOpacityRampForTimeStartOpacityEndOpacityTimeRange(time coremedia.Time, startOpacity *float32, endOpacity *float32, timeRange *coremedia.TimeRange) bool {
 	rv := objc.Call[bool](v_, objc.Sel("getOpacityRampForTime:startOpacity:endOpacity:timeRange:"), time, startOpacity, endOpacity, timeRange)
 	return rv
 }
